refactor: 음식점 리스트 관리 custom hook으로 분리
