+++ src/components/RestaurantList/index.tsx
@@ -2,9 +2,11 @@ import useWrappingContext from '../../hooks/useWrappingContext';
 import Store from '../../store';
 import RestaurantItem from '../RestaurantItem';
 import styles from './RestaurantList.module.css';
+import useRestaurantList from './hooks/useRestaurantList';
 
 function RestaurantList() {
-  const { restaurantList } = useWrappingContext(Store);
+  const { category, sortOption } = useWrappingContext(Store);
+  const restaurantList = useRestaurantList(category, sortOption);
 
   return (
     <ul className={styles.restaurantList}>
